requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
openai==0.28.0
plotly==5.3.1
numpy==1.21.4
pandas==1.3.4
python-dotenv==1.0.0
SQLAlchemy==2.0.21
requests==2.31.0
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
